+++ zowe-cli/src/jsMain/kotlin/zowe/cli/zosfiles/compare/CompareBaseHelperPrepareSpoolDescriptorResult.kt
@@ -3,5 +3,5 @@ package zowe.cli.zosfiles.compare
 external interface CompareBaseHelperPrepareSpoolDescriptorResult {
   var jobName: String
   var jobId: String
-  var spoolId: Double
+  var spoolId: Int
 }
